add RenderClient to host pixi stuff
